chore: response schema
