+++ apps/frontend/src/app/contractor/(auth)/page.tsx
@@ -2,7 +2,7 @@ export default function ContractorPage() {
     return (
         <div>
             <h1>Contractor Page</h1>
-            <p>This is the contractor page.</p>
+            <p>This is the Contractor page.</p>
         </div>
     );
 }
