+++ apps/frontend/src/app/departmentUser/(auth)/page.tsx
@@ -1,0 +1,8 @@
+export default function DepartemntUserPage() {
+    return (
+        <div>
+            <h1>Department User Page</h1>
+            <p>This is the Department User page.</p>
+        </div>
+    );
+}
