--- apps/frontend/src/app/departmentUser/page.tsx
@@ -1,8 +1,0 @@
-export default function ContractorPage() {
-    return (
-        <div>
-            <h1>Contractor Page</h1>
-            <p>This is the contractor page.</p>
-        </div>
-    );
-}
