+++ apps/frontend/src/types/response/admin/admin.ts
@@ -1,0 +1,12 @@
+import { BaseResponse } from "../baseResponse";
+
+interface AdminUser {
+    id: string;
+    username: string;
+    email: string;
+    createdAt: string;
+}
+
+export interface IAdmin extends BaseResponse {
+    data: AdminUser;
+}
