+++ apps/frontend/src/types/response/baseResponse.ts
@@ -1,0 +1,4 @@
+export interface BaseResponse {
+    success: boolean;
+    message: string;
+}
