+++ apps/frontend/src/types/schema/admin/auth.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const AdminSigninupSchema = z.object({
+    
+})
